Advent.Of.Code

Description

Advent.Of.Code

A framework for fetching inputs, submitting solutions, testing examples, and benchmarking solutions for Advent of Code.

Advent of Code

runAdvent

Arguments

:: Integer
The AoC year

-> (Day -> Part -> Maybe (Text -> Text))
Solutions

-> IO ()

The primary entry point for the framework. The solutions are expected as a function that takes the day and part and returns a Maybe (Text -> Text). That Text -> Text is used to compute the solution from the input, if it exists.

data Day

The day of a challenge.

Constructors

Day1

Day2

Day3

Day4

Day5

Day6

Day7

Day8

Day9

Day10

Day11

Day12

Day13

Day14

Day15

Day16

Day17

Day18

Day19

Day20

Day21

Day22

Day23

Day24

Day25

data Part

A given part of a problem. All Advent of Code challenges are two-parts.

You can usually get Part1 (if it is already released) with a nonsense session key, but Part2 always requires a valid session key.

Constructors

Part1

Part2