Advent.Of.Code.CLI

Description

Advent.Of.Code.CLI

Basic CLI parser for the framework.

Parsing

parser :: ParserInfo Opts

Advent of Code framework CLI option parser.

execParser :: ParserInfo a -> IO a

Run a program description.

Parse command line arguments. Display help text and exit if any parse error occurs.

Types

data Action

The possible modes of action.

Constructors

Submit
Compute solution and submit via the AoC API

ShowInput
Show the input without computing the solution

ShowOutput
Compute and print the solution without submitting

ShowPrompt
Show the prompt without computing the solution

Test FilePath
Test solutions against the examples provided in a file

Benchmark
Benchmark solutions against the input

Version
Show version and exit

data InputSource

Where to fetch the input from.

Constructors

API
Fetch from the AoC API

Stdin
Read from stdin

InputFile FilePath
Read from the specified file

data Opts

The configuration for the framework.

Constructors

MkOpts

Fields

day :: Maybe Day

part :: Maybe Part

action :: Action

inputSource :: InputSource