Advent.Of.Code.Input

Description

Advent.Of.Code.Input

Basic convenience functions for manipulating AoC inputs.

Lists

list :: Read a => Text -> [a]

Break the input into lines and parse them using a Show instance.

plainList :: Text -> [Text]

Break the input into lines.

wordsList :: ([Text] -> a) -> Text -> [a]

Break the input into lines and parse them by words using a provided function.

customList :: (Text -> a) -> Text -> [a]

Break the input into lines and parse them using a provided function.

Effectful

wordsList' :: Applicative f => ([Text] -> f a) -> Text -> f [a]

Break the input into lines and parse them by words using a provided effectful parser.

customList' :: Applicative f => (Text -> f a) -> Text -> f [a]

Break the input into lines and parse them using a provided effectful parser.

Non-Empty Lists

nonEmpty :: Read a => Text -> NonEmpty a

Break the input into lines and parse them using a Show instance.

plainNonEmpty :: Text -> NonEmpty Text

Break the input into lines.

wordsNonEmpty :: ([Text] -> a) -> Text -> NonEmpty a

Break the input into lines and parse them by words using a provided function.

customNonEmpty :: (Text -> a) -> Text -> NonEmpty a

Break the input into lines and parse them using a provided function.

Effectful

wordsNonEmpty' :: Applicative f => ([Text] -> f a) -> Text -> f (NonEmpty a)

Break the input into lines and parse them by words using a provided effectful parser.

customNonEmpty' :: Applicative f => (Text -> f a) -> Text -> f (NonEmpty a)

Break the input into lines and parse them using a provided effectful parser.

Miscellaneous

readt :: Read a => Text -> a

Parse a Text value using a Show instance.