Advent.Of.Code.Test.Types

Description

Advent.Of.Code.Test.Types

Types for constructing tests of solutions.

Documentation

data Test

A single test, consisting of a given input and expected output.

Constructors

Test

Fields

input :: Text

output :: Text

type Tests = Map (Day, Part) [Test]

A set of tests, organized by day and part.